backend/requirements.txt
# Core ML and Data Science
numpy>=1.21.0
pandas>=1.5.0
scikit-learn>=1.1.0
joblib>=1.2.0

# FastAPI Stack (for API development)
fastapi>=0.95.0
uvicorn>=0.20.0
pydantic>=2.0.0

# Streamlit (for web interfaces)
streamlit>=1.25.0

# HTTP Client (for API examples)
requests>=2.28.0

# OpenAI SDK (for LLM interactions)
openai>=1.10.0

# PPTX generation (for creating PowerPoint files)
python-pptx>=0.6.23

# Environment Variables Management (for loading .env files)
python-dotenv>=1.2.0

# ── Installation ─────────────────────────────────────────────────────────────
# pip install -r requirements_minimal.txt 